moved the routes file and updated file paths
