+++ db/db.json
@@ -1,1 +1,1 @@
-[{"id":"592c5f11-b395-42c4-8dbb-841e71963835","title":"asdf","text":"asdf"}]
+[]
